% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnc_RNARNA_scanner_output.R
\docType{data}
\name{lnc_RNARNA_scanner_output}
\alias{lnc_RNARNA_scanner_output}
\title{lnc_RNARNA_scanner_output}
\format{
a list of data frames
\describe{
}
}
\source{
lnc_RNARNA_scanner function
}
\usage{
lnc_RNARNA_scanner_output
}
\description{
example output from lnc_RNARNA_scanner function, tables are truncated to reduced size of the file
}
\examples{
\dontrun{
 data(lnc_RNARNA_scanner_output)
}



}
\references{
RNARNA.db creators: Terai G, Iwakiri J, Kameda T, Hamada M, Asai K. Comprehensive prediction of lncRNA-RNA interactions in human transcriptome. BMC Genomics. 2016;17 Suppl 1(Suppl 1):12. Published 2016 Jan 11. doi:10.1186/s12864-015-2307-5
http://rtools.cbrc.jp/cgi-bin/RNARNA/index.pl#'
}
\keyword{datasets}
