% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topmiRNA_toptarget_output.R
\docType{data}
\name{topmiRNA_toptarget_output}
\alias{topmiRNA_toptarget_output}
\title{topmiRNA_toptarget_output}
\format{
A list of data frames
\describe{
}
}
\source{
topmiRNA_toptarget function
}
\usage{
topmiRNA_toptarget_output
}
\description{
example output from topmiRNA_toptarget function
}
\examples{
\dontrun{
 data(topmiRNA_toptarget_output)
}



}
\references{
Ru Y, Kechris KJ, Tabakoff B, Hoffman P, Radcliffe RA, Bowler R, Mahaffey S, Rossi S, Calin GA, Bemis L, Theodorescu D (2014). “The multiMiR R package and database: integration of microRNA–target interactions along with their disease and drug associations.” Nucleic Acids Research, 42(17), e133. doi: 10.1093/nar/gku631, http://dx.doi.org/10.1093/nar/gku631.
Ru Y, Mulvahill M, Mahaffey S, Kechris K. multiMiR: Integration of multiple microRNA-target databases with their disease and drug associations. https://github.com/KechrisLab/multiMiR.
}
\keyword{datasets}
